cmd/conn: define Marshal on Marshallable
