+++ cmd/conn/chunker.go
@@ -5,7 +5,7 @@ import "github.com/WatchBeam/rtmp/chunk"
 // Chunker is a functional interface designed to turn ConnSendable types into
 // RTMP `*chunk.Chunk`s.
 type Chunker interface {
-	// Chunk marshals a ConnSendable into the *chunk.Chunk type, using the
+	// Chunk marshals a Marshallable into the *chunk.Chunk type, using the
 	// protocol defined in the RTMP specification.
-	Chunk(s Sendable) (*chunk.Chunk, error)
+	Chunk(m Marshallable) (*chunk.Chunk, error)
 }
